@@ -8,8 +8,8 @@ class FileTest extends \PHPUnit\Framework\TestCase
 {
     public function teardown(): void
     {
-        if (file_exists(SABRE_TEMPDIR.'/filebackend')) {
-            unlink(SABRE_TEMPDIR.'/filebackend');
+        if (file_exists(\Sabre\TestUtil::SABRE_TEMPDIR.'/filebackend')) {
+            unlink(\Sabre\TestUtil::SABRE_TEMPDIR.'/filebackend');
         }
     }
 
@@ -22,15 +22,15 @@ class FileTest extends \PHPUnit\Framework\TestCase
     public function testLoadFileBroken()
     {
         $this->expectException('Sabre\DAV\Exception');
-        file_put_contents(SABRE_TEMPDIR.'/backend', 'user:realm:hash');
-        $file = new File(SABRE_TEMPDIR.'/backend');
+        file_put_contents(\Sabre\TestUtil::SABRE_TEMPDIR.'/backend', 'user:realm:hash');
+        $file = new File(\Sabre\TestUtil::SABRE_TEMPDIR.'/backend');
     }
 
     public function testLoadFile()
     {
-        file_put_contents(SABRE_TEMPDIR.'/backend', 'user:realm:'.md5('user:realm:password'));
+        file_put_contents(\Sabre\TestUtil::SABRE_TEMPDIR.'/backend', 'user:realm:'.md5('user:realm:password'));
         $file = new File();
-        $file->loadFile(SABRE_TEMPDIR.'/backend');
+        $file->loadFile(\Sabre\TestUtil::SABRE_TEMPDIR.'/backend');
 
         self::assertFalse($file->getDigestHash('realm', 'blabla'));
         self::assertEquals(md5('user:realm:password'), $file->getDigestHash('realm', 'user'));
