@@ -12,8 +12,8 @@ class GuessContentTypeTest extends DAV\AbstractServer
     {
         parent::setUp();
         \Sabre\TestUtil::clearTempDir();
-        file_put_contents(SABRE_TEMPDIR.'/somefile.jpg', 'blabla');
-        file_put_contents(SABRE_TEMPDIR.'/somefile.hoi', 'blabla');
+        file_put_contents(\Sabre\TestUtil::SABRE_TEMPDIR.'/somefile.jpg', 'blabla');
+        file_put_contents(\Sabre\TestUtil::SABRE_TEMPDIR.'/somefile.hoi', 'blabla');
     }
 
     public function tearDown(): void
