@@ -8,13 +8,13 @@ class DirectoryTest extends \PHPUnit\Framework\TestCase
 {
     public function create()
     {
-        return new Directory(SABRE_TEMPDIR);
+        return new Directory(\Sabre\TestUtil::SABRE_TEMPDIR);
     }
 
     public function testCreate()
     {
         $dir = $this->create();
-        self::assertEquals(basename(SABRE_TEMPDIR), $dir->getName());
+        self::assertEquals(basename(\Sabre\TestUtil::SABRE_TEMPDIR), $dir->getName());
     }
 
     public function testChildExistDot()
