@@ -12,7 +12,7 @@ class Plugin2Test extends \Sabre\DAVServerTest
 
     public function setUpTree()
     {
-        $this->tree = new \Sabre\DAV\FS\Directory(SABRE_TEMPDIR);
+        $this->tree = new \Sabre\DAV\FS\Directory(\Sabre\TestUtil::SABRE_TEMPDIR);
     }
 
     public function teardown(): void
