@@ -4,23 +4,25 @@ declare(strict_types=1);
 
 namespace Sabre\DAVACL\FS;
 
+use Sabre\TestUtil;
+
 class CollectionTest extends FileTest
 {
     public function setup(): void
     {
         parent::setup();
-        $this->path = SABRE_TEMPDIR;
+        $this->path = TestUtil::SABRE_TEMPDIR;
         $this->sut = new Collection($this->path, $this->acl, $this->owner);
     }
 
     public function teardown(): void
     {
-        \Sabre\TestUtil::clearTempDir();
+        TestUtil::clearTempDir();
     }
 
     public function testGetChildFile()
     {
-        file_put_contents(SABRE_TEMPDIR.'/file.txt', 'hello');
+        file_put_contents(TestUtil::SABRE_TEMPDIR.'/file.txt', 'hello');
         $child = $this->sut->getChild('file.txt');
         self::assertInstanceOf('Sabre\\DAVACL\\FS\\File', $child);
 
@@ -31,7 +33,7 @@ class CollectionTest extends FileTest
 
     public function testGetChildDirectory()
     {
-        mkdir(SABRE_TEMPDIR.'/dir');
+        mkdir(TestUtil::SABRE_TEMPDIR.'/dir');
         $child = $this->sut->getChild('dir');
         self::assertInstanceOf('Sabre\\DAVACL\\FS\\Collection', $child);
 
